parametrize set_axis tests (#37619)
